# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenAPI natation project backend documentation",
        "description": "backend is used by smartphone application and desktop front",
        "contact": {
            "email": "**"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/programme/{programme}/exercice/{exercise}": {
            "put": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Validate and mark an exercise as finished",
                "description": "Endpoint to validate exercise completion with optional performance data",
                "operationId": "validateExercise",
                "parameters": [
                    {
                        "name": "programme",
                        "in": "path",
                        "description": "Program ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "exercise",
                        "in": "path",
                        "description": "Exercise ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "ID of the user completing the exercise",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "duration_completed": {
                                        "description": "Actual duration in seconds",
                                        "type": "integer",
                                        "example": 300
                                    },
                                    "distance_completed": {
                                        "description": "Actual distance completed",
                                        "type": "number",
                                        "format": "float",
                                        "example": 1500.5
                                    },
                                    "repetitions_completed": {
                                        "description": "Actual repetitions completed",
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "weight_used": {
                                        "description": "Weight used during exercise",
                                        "type": "number",
                                        "format": "float",
                                        "example": 75.5
                                    },
                                    "notes": {
                                        "description": "Additional notes",
                                        "type": "string",
                                        "example": "Exercise completed successfully"
                                    },
                                    "gps_data": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "latitude": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 48.856614
                                                },
                                                "longitude": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 2.352222
                                                },
                                                "timestamp": {
                                                    "type": "string",
                                                    "example": "2024-01-01T12:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exercise marked as finished successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Exercise completed successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "program_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "exercise_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "finished_at": {
                                                    "type": "string",
                                                    "example": "2024-01-01T12:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Program or exercise not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Exercise not found in this program"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid input"
                                        },
                                        "details": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
    },
...
}